import 2020-08-10.7z

+++ src/vm/fm7/fm7.cpp
@@ -1268,7 +1268,7 @@ bool VM::process_state(FILEIO* state_fio, bool loading)
 	state_fio->StateValue(connect_1Mfdc);
 	for(DEVICE* device = first_device; device; device = device->next_device) {
 		const char *name = typeid(*device).name() + 6; // skip "class "
-		int len = strlen(name);
+		int len = (int)strlen(name);
 		
 		if(!state_fio->StateCheckInt32(len)) {
 			return false;
